lesson_02/job1/requirements.txt
Flask==2.2.2
requests==2.28.1
black==22.12.0
fastavro==1.7.0
python-dotenv==0.21.0
